pyproject.toml
[project]
name = "bolt-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "fpdf2>=2.8.3",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "langchain-openai>=0.3.23",
    "langgraph>=0.4.8",
    "uvicorn>=0.34.3",
]
